Implement auto-flushing for idle consoles
Every 30 sec if nothing has been written but the console is active a '.' is written to keep the TCP connection alive. Azure, for example, terminates TCP connections with no data for > 4 mins. So this gets around that and provides an alive sign to anyone watching a long running task.

As well fix writing HTML to the end of TextWebConsole after it's disposed.